Add experimental obpf/lobby support

+++ platforms/build-android.sh
@@ -2,6 +2,9 @@
 
 set -e
 
+## options: "smart, complete_rebuild"
+export COMPILE_TYPE="smart"
+
 mkdir -p toolchains
 
 export NDK_VER_DOWNLOAD="r26c"
@@ -15,8 +18,10 @@ if [ ! -d "$BASE_PATH" ]; then
 
     cd toolchains
 
-    wget -q "https://dl.google.com/android/repository/android-ndk-$NDK_VER_DOWNLOAD-linux.zip"
+    if [ ! -e "android-ndk-$NDK_VER_DOWNLOAD-linux.zip" ]; then
 
+        wget -q "https://dl.google.com/android/repository/android-ndk-$NDK_VER_DOWNLOAD-linux.zip"
+    fi
     unzip -q "android-ndk-$NDK_VER_DOWNLOAD-linux.zip"
 
     cd ..
@@ -63,7 +68,16 @@ elif [ "$#" -eq 1 ]; then
 
     ARCH_KEYS_INDEX=("$FOUND")
 else
-    echo "Too many arguemtns given, expected at most 1"
+    echo "Too many argumetns given, expected at most 1"
+    exit 1
+fi
+
+if [ "$COMPILE_TYPE" == "smart" ]; then
+    : # noop
+elif [ "$COMPILE_TYPE" == "complete_rebuild" ]; then
+    : # noop
+else
+    echo "Invalid COMPILE_TYPE, expected: 'smart' or 'complete_rebuild'"
     exit 1
 fi
 
@@ -74,14 +88,14 @@ for INDEX in "${ARCH_KEYS_INDEX[@]}"; do
 
     BITNESS=$(echo "$RAW_JSON" | jq -M -r -c '."bitness"') || true
     ARCH=$(echo "$RAW_JSON" | jq -M -r -c '."arch"')
-    ARM_VERSION=$(echo "$RAW_JSON" | jq -M -r -c '."proc"' | tr -d "-")
+    ARCH_VERSION=$(echo "$RAW_JSON" | jq -M -r -c '."proc"' | tr -d "-")
     ARM_NAME_TRIPLE=$(echo "$RAW_JSON" | jq -M -r -c '."triple"')
     ARM_TARGET_ARCH=$KEY
     ARM_TRIPLE=$ARM_NAME_TRIPLE$SDK_VERSION
     ARM_COMPILER_TRIPLE=$(echo "$RAW_JSON" | jq -M -r -c '."llvm_triple"')
-    ARM_TOOL_TRIPLE=$(echo "$ARM_NAME_TRIPLE$SDK_VERSION" | sed s/$ARCH/$ARM_VERSION/)
+    ARM_TOOL_TRIPLE=$(echo "$ARM_NAME_TRIPLE$SDK_VERSION" | sed s/$ARCH/$ARCH_VERSION/)
 
-    export SYM_LINK_PATH=sysroot_sym-$ARM_VERSION
+    export SYM_LINK_PATH=sysroot_sym-$ARCH_VERSION
 
     export HOST_ROOT="$BASE_PATH/toolchains/llvm/prebuilt/linux-x86_64"
     export SYS_ROOT="${HOST_ROOT}/$SYM_LINK_PATH"
@@ -93,32 +107,36 @@ for INDEX in "${ARCH_KEYS_INDEX[@]}"; do
 
     export LIBRARY_PATH="$SYS_ROOT/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION"
 
-    LAST_DIR=$PWD
+    if [ "$COMPILE_TYPE" == "complete_rebuild" ] || ! [ -e "$SYS_ROOT" ]; then
 
-    if [ -d "${SYS_ROOT:?}/" ]; then
+        LAST_DIR=$PWD
 
-        rm -rf "${SYS_ROOT:?}/"
+        if [ -d "${SYS_ROOT:?}/" ]; then
+
+            rm -rf "${SYS_ROOT:?}/"
+        fi
+
+        mkdir -p "${SYS_ROOT:?}/usr/lib"
+
+        cd "${SYS_ROOT:?}/usr/"
+
+        ln -s "$HOST_ROOT/sysroot/usr/local" "${SYS_ROOT:?}/usr/"
+
+        ln -s "$HOST_ROOT/sysroot/usr/include" "${SYS_ROOT:?}/usr/"
+
+        find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/" -maxdepth 1 -name "*.so" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
+
+        find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/" -maxdepth 1 -name "*.a" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
+
+        find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.a" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
+
+        find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.so" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
+
+        find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.o" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
+
+        cd "$LAST_DIR"
+
     fi
-
-    mkdir -p "${SYS_ROOT:?}/usr/lib"
-
-    cd "${SYS_ROOT:?}/usr/"
-
-    ln -s "$HOST_ROOT/sysroot/usr/local" "${SYS_ROOT:?}/usr/"
-
-    ln -s "$HOST_ROOT/sysroot/usr/include" "${SYS_ROOT:?}/usr/"
-
-    find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/" -maxdepth 1 -name "*.so" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
-
-    find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/" -maxdepth 1 -name "*.a" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
-
-    find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.a" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
-
-    find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.so" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
-
-    find "$HOST_ROOT/sysroot/usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION/" -maxdepth 1 -name "*.o" -exec ln -s "{}" "${SYS_ROOT:?}/usr/lib/" \;
-
-    cd "$LAST_DIR"
 
     export BUILD_DIR="build-$ARM_TARGET_ARCH"
 
@@ -131,57 +149,111 @@ for INDEX in "${ARCH_KEYS_INDEX[@]}"; do
     export STRIP=llvm-strip
     unset PKG_CONFIG
 
+    ## BUILD dependencies not buildable with meson (to complicated to port)
+
     ## build mpg123 with autotools (meson port is to much work atm, for this feature)
 
-    LAST_DIR=$PWD
+    LAST_DIR="$PWD"
 
     cd "$SYS_ROOT"
 
-    mkdir -p "build"
+    BUILD_DIR_MPG123="build-mpg123"
 
-    cd build
+    BUILD_MPG123_FILE="$SYS_ROOT/$BUILD_DIR_MPG123/build_succesfull.meta"
 
-    wget -q "https://www.mpg123.de/download/mpg123-1.31.3.tar.bz2"
+    if [ "$COMPILE_TYPE" == "complete_rebuild" ] || ! [ -e "$BUILD_MPG123_FILE" ]; then
 
-    tar -xf "mpg123-1.31.3.tar.bz2"
+        mkdir -p "$BUILD_DIR_MPG123"
 
-    cd "mpg123-1.31.3"
+        cd "$BUILD_DIR_MPG123"
 
-    BUILDYSTEM="cmake"
+        wget -q "https://www.mpg123.de/download/mpg123-1.32.4.tar.bz2"
 
-    if [ $BUILDYSTEM = "autotools" ]; then
+        tar -xf "mpg123-1.32.4.tar.bz2"
 
-        ./configure --prefix="$SYS_ROOT/usr" --oldincludedir="$SYS_ROOT/usr/include" --host="$ARM_NAME_TRIPLE" --with-sysroot="$SYS_ROOT" --with-audio="dummy"
+        cd "mpg123-1.32.4"
 
-        make
+        BUILDYSTEM="cmake"
 
-        make install
+        if [ $BUILDYSTEM = "autotools" ]; then
 
-    else
+            ./configure --prefix="$SYS_ROOT/usr" --oldincludedir="$SYS_ROOT/usr/include" --host="$ARM_NAME_TRIPLE" --with-sysroot="$SYS_ROOT" --with-audio="dummy"
 
-        cd ports/cmake/
+            make
 
-        mkdir -p build
-
-        cd build
-
-        if [ "$ARM_VERSION" = "i686" ]; then
-            #cmake .. -DCMAKE_TOOLCHAIN_FILE=linux_i686.toolchain.cmake --install-prefix "$SYS_ROOT/usr" "-DCMAKE_SYSROOT=$SYS_ROOT" -DOUTPUT_MODULES=dummy -DCMAKE_POSITION_INDEPENDENT_CODE=ON
-            # cmake --build .
-
-            # cmake --install .
-            : # nop, for bash syntax
+            make install
 
         else
-            cmake .. --install-prefix "$SYS_ROOT/usr" "-DCMAKE_SYSROOT=$SYS_ROOT" -DOUTPUT_MODULES=dummy -DCMAKE_POSITION_INDEPENDENT_CODE=ON
-            cmake --build .
 
-            cmake --install .
+            cd ports/cmake/
+
+            BUILD_DIR_MPG123="build-mpg123"
+
+            mkdir -p "$BUILD_DIR_MPG123"
+
+            cd "$BUILD_DIR_MPG123"
+
+            if [ "$ARCH_VERSION" = "i686" ]; then
+                #cmake .. -DCMAKE_TOOLCHAIN_FILE=linux_i686.toolchain.cmake --install-prefix "$SYS_ROOT/usr" "-DCMAKE_SYSROOT=$SYS_ROOT" -DOUTPUT_MODULES=dummy -DCMAKE_POSITION_INDEPENDENT_CODE=ON
+                # cmake --build .
+
+                # cmake --install .
+                : # nop, for bash syntax
+
+            else
+                cmake .. --install-prefix "$SYS_ROOT/usr" "-DCMAKE_SYSROOT=$SYS_ROOT" -DOUTPUT_MODULES=dummy -DCMAKE_POSITION_INDEPENDENT_CODE=ON
+                cmake --build .
+
+                cmake --install .
+            fi
+
         fi
+
+        touch "$BUILD_MPG123_FILE"
 
     fi
 
     cd "$LAST_DIR"
+
+    ## build openssl with make (meson port is to much work atm, for this feature)
+
+    LAST_DIR="$PWD"
+
+    cd "$SYS_ROOT"
+
+    BUILD_DIR_OPENSSL="build-openssl"
+
+    BUILD_OPENSSL_FILE="$SYS_ROOT/$BUILD_DIR_OPENSSL/build_succesfull.meta"
+
+    if [ "$COMPILE_TYPE" == "complete_rebuild" ] || ! [ -e "$BUILD_OPENSSL_FILE" ]; then
+
+        mkdir -p "$BUILD_DIR_OPENSSL"
+
+        cd "$BUILD_DIR_OPENSSL"
+
+        wget -q "https://www.openssl.org/source/openssl-3.0.13.tar.gz"
+
+        tar -xzf "openssl-3.0.13.tar.gz"
+
+        cd "openssl-3.0.13"
+
+        OPENSSL_TARGET_ARCH="android-$ARCH"
+
+        export ANDROID_NDK_ROOT="$ANDROID_NDK_HOME"
+
+        ./Configure --prefix="$SYS_ROOT/usr" no-tests no-shared "$OPENSSL_TARGET_ARCH" "-D__ANDROID_API__=$SDK_VERSION"
+
+        make -j build_sw
+
+        make install_sw
+
+        touch "$BUILD_OPENSSL_FILE"
+
+    fi
+
+    cd "$LAST_DIR"
+
+    ## END of manual build of dependencies
 
     MESON_CPU_FAMILY=$ARCH
 
@@ -194,7 +266,7 @@ for INDEX in "${ARCH_KEYS_INDEX[@]}"; do
 [host_machine]
 system = 'android'
 cpu_family = '$MESON_CPU_FAMILY'
-cpu = '$ARM_VERSION'
+cpu = '$ARCH_VERSION'
 endian = 'little'
 
 [constants]
@@ -254,15 +326,23 @@ EOF
 
     export LIBRARY_PATH="$LIBRARY_PATH:usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION:$LIB_PATH"
 
-    meson setup "$BUILD_DIR" \
-        "--prefix=$SYS_ROOT" \
-        "--includedir=$INC_PATH" \
-        "--libdir=usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION" \
-        --cross-file "./platforms/crossbuild-android-$ARM_TARGET_ARCH.ini" \
-        -Dsdl2:use_hidapi=disabled \
-        -Dcpp_args=-DAUDIO_PREFER_MP3
+    if [ "$COMPILE_TYPE" == "complete_rebuild" ] || [ ! -e "$BUILD_DIR" ]; then
+
+        meson setup "$BUILD_DIR" \
+            "--prefix=$SYS_ROOT" \
+            "--wipe" \
+            "--includedir=$INC_PATH" \
+            "--libdir=usr/lib/$ARM_NAME_TRIPLE/$SDK_VERSION" \
+            --cross-file "./platforms/crossbuild-android-$ARM_TARGET_ARCH.ini" \
+            -Dbuildtype=release \
+            -Dsdl2:use_hidapi=disabled \
+            -Dcpp_args=-DAUDIO_PREFER_MP3
+
+    fi
 
     meson compile -C "$BUILD_DIR"
+
+    echo -e "Sucesfully build for andropid platform ${ARCH}\n"
 
 done
